rename: 여행 관련 api 폴더 및 import 정리
