+++ src/@types/travel.ts
@@ -89,3 +89,8 @@ export interface TravelList {
   maxTravelMateCount: number;
   currentTravelMateCount: number;
 }
+
+export interface MyTravel {
+  travels: TravelList[];
+  total: number;
+}
